% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{ambhasGW}
\alias{ambhasGW}
\title{Ground Water Modelling}
\usage{
ambhasGW(input.file)
}
\arguments{
\item{input.file}{yml file having input parameters information}
}
\description{
\code{ambhasGW} Computes the groundwater head for each time step 
in raster format
}
\examples{

# Create necessary input file to make dummy run
# Dummy directory to run
dummy.dir <- '~'

# Make dummy run
create_inputs(dummy.dir)
input.file <- file.path(dummy.dir , 'input/parameters.yml')
ambhasGW(input.file)

}
