% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_inputs.R
\name{create_inputs}
\alias{create_inputs}
\title{Create dummy input data}
\usage{
create_inputs(dummy.dir)
}
\arguments{
\item{dummy.dir}{Dummy directory to input and output data}
}
\description{
Creates dummy data and corresponding yml file to run the model.
}
\examples{



# Creates input data and yml file
create_inputs(tempdir())


}
