% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter_definitions.R
\name{parameter_definition}
\alias{parameter_definition}
\title{Parameters definition}
\usage{
parameter_definition(par)
}
\arguments{
\item{par}{Parameter name}
}
\description{
Prints the parameters definition of given parameter
}
\examples{
 
parameter_definition('hini')

}
